requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Django==4.1.3
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
pycparser==2.21
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12